// src/app/precisa/ventas/Interes/model/Dtointeres.ts
export class DtoInteres{
    constructor() {
    }

    num?:                 number;
    CompaniaCodigo?:      string;
    DescripcionCorta?:    string;
    IdSede?:              number;
    SedDescripcion?:      string;
    NomProyecto?:         string;
    NomManzana?:          string;
    DesCondicion?:        string;
    DesSituacion?:        string;
    IdLote?:              number;
    DesLote?:             string;
    Situacion?:           number;
    IdReserva?:           number;
    IdVendedor?:          number;
    IdProyecto?:          number;
    IdManzana?:           number;
    TipoDocumento?:       string;
    Documento?:           string;
    Cliente?:             string;
    TipoDocConyuge?:      string;
    DocConyuge?:          string;
    Conyuge?:             string;
    TipoDocVendedor?:     string;
    DocVendedor?:         string;
    Vendedor?:            string;
    DesUbicacion?:        string;
    DesTipoPago?:         string;
    DesBanco?:            string;
    IdInteres?:           number;
    IdLetra?:             number;
    IdContrato?:          number;
    IdCliente?:           number;
    FechaRegistro?:       Date;
    Codigo?:              string;
    Nombre?:              string;
    Observacion?:         string;
    FechaVencimiento?:    Date;
    FechaPago?:           Date;
    MonedaCodigo?:        string;
    MontoLetra?:          number;
    DiasMora?:            number;
    TotalMora?:           number;
    Descuento?:           number;
    MontoPago?:           number;
    Estado?:              number;
    UsuarioCreacion?:     string;
    UsuarioModificacion?: string;
    FechaCreacion?:       Date;
    FechaModificacion?:   Date;
    IpCreacion?:          string;
    IpModificacion?:      string; 

    TipoMotivo?:          number;
    IdComprobante?:       number;
    FormaPago?:           string;
    Banco?:               string;
    CuentaBancaria?:      string;
    TipoDescuento?:       number;
    InteresPadre?:        number;
    DesEstado?:           string;
    MontoTotal?:          number;
}