// src/app/precisa/ventas/Control/model/dtoLetra.ts
export class DtoLetra {
    constructor(){

    }
    CompaniaCodigo:      string;
    DescripcionCorta:    string;
    IdSede:              number;
    SedDescripcion:      string;
    NomProyecto:         string;
    NomManzana:          string;
    DesCondicion:        string;
    DesSituacion:        string;
    IdLote:              number;
    DesLote:             string;
    Situacion:           number;
    IdReserva:           number;
    IdProyecto:          number;
    IdManzana:           number;
    TipoDocumento:       string;
    Documento:           string;
    Cliente:             string;
    TipoDocConyuge:      string;
    DocConyuge:          string;
    Conyuge:             string;
    IdVendedor:          string;
    TipoDocVendedor:     string;
    DocVendedor:         string;
    Vendedor:            string;
    DesUbicacion:        string;
    TipoPago:            string;
    Banco:               string;
    IdLetra:             number;
    IdContrato:          number;
    IdCliente:           number;
    Codigo:              string;
    codigocomponente:    string;
    Nombre:              string;
    FechaInicio:         string;
    FechaVencimiento:    Date;
    FechaPago:           string;
    MonedaCodigo:        string;

    Interes:             number;
    diasAtrasados:       number;

    Amortizacion:        string;
    Cuota:               number;
    MontoAfecto:         number;
    MontoImpuesto:       string;
    MontoTotal:          number;
    IdComprobante:       string;
    Observacion:         string;
    Estado:              number;
    UsuarioCreacion:     string;
    UsuarioModificacion: string;
    FechaCreacion:       Date;
    FechaModificacion:   Date;
    IpCreacion:          string;
    IpModificacion:      string;
    DesEstado:           string;
    Link_pdf:            string;
    num:                 number;
    flatApro:            number;
}